Inicio sesion, registrar terminado a falta de hacer restricciones y notas medio acabado

+++ Proyecto-Kotlin-T1/app/src/main/java/com/example/proyecto_kotlin/modelos/Nota.kt
@@ -2,11 +2,11 @@ package com.example.proyecto_kotlin.modelos
 
 import java.time.LocalDate
 
-data class Nota (
-    var id: Int? = null,
+data class Nota(
+    var id: Long ?= null,
     var titulo: String,
     var texto: String,
     var estado: String,
-    var fechaCreacion : LocalDate,
+    var fechaCreacion: LocalDate,
     var idUsuario: Long
 )
